requirements.txt
streamlit
google-generativeai
edge-tts
fastapi
uvicorn
python-multipart
Pillow
extra-streamlit-components
